Update tips.md
Sorry for the delay. I think this should cover the discussion tips with an extra one for the `TODO` tip. I wasn't sure if I should add that you can navigate `TODO`s as I don't use it due to the enormous amount of them in our codebase so I've no idea how ergonomic that is compared to the alternatives.