docs(recipes): Add recipe for enabling supertab behavior in blink.cmp
This recipe sets up blink.cmp's bindings such that <Tab> will navigate
down a list of completions, and <S-Tab> will navigate up that list.
Enter will submit the selected completion and drop the user into the
completion's parameter(s) or directly after the inserted text, whichever
applies to that completion.

For reference, I had to dig for close to an hour to figure out how to do
this.

The following reddit thread is a collection of similar frustrations, as
well as multiple outdated ways to achieve this:
https://www.reddit.com/r/neovim/comments/1hfotru/nvimcmp_super_tab_in_blink/

An incomplete solution is here:
https://github.com/LazyVim/LazyVim/discussions/250#discussioncomment-12461707

But the preset must be set to "enter", as noted here:
https://github.com/Saghen/blink.cmp/discussions/576#discussioncomment-11599108

I plainly cannot find where this is documented within blink.cmp's docs.